new version

+++ app/Http/Controllers/Admin/ProfileController.php
@@ -51,7 +51,12 @@ class ProfileController extends \App\Http\Controllers\Controller
 
         if ($oldName != $user->name || $oldLastName != $user->lastname || $oldEmail != $user->email || $oldRole != $user->role || $request->hasFile('profile_img')) {
             if ($request->file('profile_img') != null) {
-                Storage::delete($oldProfile);
+                if ($oldProfile == null) {
+                    $this->processFile($user->id, $user);
+                }
+                if ($oldProfile != null) {
+                    Storage::delete($oldProfile);
+                }
                 $this->processFile($user->id, $user);
             } else {
                 $user->profile_img = $oldProfile;
@@ -79,7 +84,7 @@ class ProfileController extends \App\Http\Controllers\Controller
             return false;
         }
         $fileName = $user->name . '_' . $id . '.' . $file->extension();
-        $file = $file->storeAs('public/' . env('PROFILE_IMG_DIR'), $fileName);
+        $file = $file->storeAs(env('PROFILE_IMG_DIR'), $fileName, 'public');
         $filePath = public_path('storage/' . env('PROFILE_IMG_DIR') . '/' . $fileName);
         $this->createThumbnail($filePath);
         $user->profile_img = env('PROFILE_IMG_DIR') . $fileName;
